+++ targets/elecrow-rp2350.json
@@ -1,6 +1,6 @@
 {
     "inherits": ["rp2350"],
-    "build-tags": ["elecrow_rp2350"],
+    "build-tags": ["elecrow_rp2350", "rp2350a"],
     "serial-port": ["2e8a:000f"],
     "default-stack-size": 8192,
     "ldflags": [
