+++ targets/pico2.json
@@ -2,7 +2,7 @@
     "inherits": [
         "rp2350"
     ],
-    "build-tags": ["pico2"],
+    "build-tags": ["pico2", "rp2350a"],
     "serial-port": ["2e8a:000A"],
     "default-stack-size": 8192
 }
